+++ crates/miden-objects/src/testing/account.rs
@@ -79,7 +79,7 @@ impl Account {
         };
 
         // construct nft tree
-        let nft_storage_map = StorageMap::with_entries(entries);
+        let nft_storage_map = StorageMap::with_entries(entries).unwrap();
 
         let account_id = AccountId::try_from(account_id).unwrap();