+++ crates/miden-tx-batch-prover/src/errors.rs
@@ -3,7 +3,7 @@ use miden_tx::TransactionVerifierError;
 use thiserror::Error;
 
 #[derive(Debug, Error)]
-pub enum BatchProveError {
+pub enum ProvenBatchError {
     #[error("failed to verify transaction {transaction_id} in transaction batch")]
     TransactionVerificationFailed {
         transaction_id: TransactionId,