+++ crates/miden-tx/src/testing/mock_chain/mod.rs
@@ -461,18 +461,16 @@ impl MockChain {
             id,
             account_updates,
             input_notes,
-            output_notes_tree,
             output_notes,
             batch_expiration_block_num,
         ) = proposed_batch.into_parts();
 
-        ProvenBatch::new(
+        ProvenBatch::new_unchecked(
             id,
             block_header.hash(),
             block_header.block_num(),
             account_updates,
             input_notes,
-            output_notes_tree,
             output_notes,
             batch_expiration_block_num,
         )